more tests and cleanup
