+++ lib/app.js
@@ -10,7 +10,6 @@ const _ = require('./lo')
 const Context = require('./context')
 const { exec } = require('./run')
 const mu = require('./utils')
-const pu = require('./pu')
 const Request = require('./request')
 const Response = require('./response')
 
@@ -97,7 +96,7 @@ class Mali extends Emitter {
       proto = grpc.loadPackageDefinition(pd)
     }
 
-    const data = pu.getServiceDefintions(proto)
+    const data = mu.getServiceDefintions(proto)
 
     if (!name) {
       name = _.keys(data)
@@ -368,10 +367,6 @@ class Mali extends Emitter {
 
     server.tryShutdownAsync = util.promisify(server.tryShutdown)
 
-    console.log('===========')
-    console.dir(this.data, { depth: 3, colors: true })
-    console.log('===========')
-
     _.forOwn(this.data, (sd, sn) => {
       const composed = {}
       const { methods, handlers } = sd
@@ -387,9 +382,6 @@ class Mali extends Emitter {
 
           composed[shortComposedKey] = this.callback(methods[k], v)
         })
-
-        console.log(sn)
-        console.log(composed)
 
         server.addService(sd.service, composed)
       }
