--- protos/messages/hello_reply.proto
@@ -1,7 +1,0 @@
-syntax = "proto3";
-
-package helloworld;
-
-message HelloReply {
-  string message = 1;
-}
