--- protos/messages/hello_request.proto
@@ -1,7 +1,0 @@
-syntax = "proto3";
-
-package helloworld;
-
-message HelloRequest {
-  string name = 1;
-}
