+++ test/protos/multipkg2.proto
